day_15_higher_order_func/pyproject.toml
[tool.poetry]
name = "higher_order_functions"
version = "0.1.0"
description = ""
authors = ["ArjanCodes <support@arjancodes.com>"]

[tool.poetry.dependencies]
python = "^3.11.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
]
